Feature/221 move first pipe (#375)
* Fix moving while inserting code in monaco

* Fix the same statechange bug as with settings but at search

* Add possability to add, move and remove firstPipe
The firtPipe will always be changed if the connection is moved or if a new one is created. There can only be one firstPipe after all.
I'm not realy proud of the code, while it works and is understandable, I don't like how it doesn't reuse existing code. Hopefully this can be changed when we're going to refactor the xml-to-flow-structure worker. The pipeline doesn't have an id or node at the moment, this could change

* Cleanup code and process feedback

* Fix addAttribute in options